% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{AIC}
\alias{AIC}
\title{Model criterion function}
\usage{
AIC(object)
}
\arguments{
\item{object}{the object to be passed to the concrete class constructor \code{BIC}}
}
\description{
A generic S3 function the Schwarz Information Criterion
}
