% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{AIC.regcd}
\alias{AIC.regcd}
\title{Schwarz Information Criterion}
\usage{
\method{AIC}{regcd}(object)
}
\arguments{
\item{object}{object returned as output from \link{reg_cd}}
}
\value{
BIC value
}
\description{
Compute the Schwarz Information Criterion for the general class of models
}
