% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{BIC.lmcd}
\alias{BIC.lmcd}
\title{Schwarz Information Criterion - lmcd class}
\usage{
\method{BIC}{lmcd}(object)
}
\arguments{
\item{object}{object returned as output from \link{reg_cd}}
}
\value{
BIC value
}
\description{
A class specific implementation of the Schwarz Information Criterion for "lmcd" class
}
