% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{c_crit}
\alias{c_crit}
\title{Concrete class constructor}
\usage{
c_crit(Y, lc, eta = NULL, klass = "lm")
}
\arguments{
\item{Y}{vector of the outcome variable}

\item{lc}{linear combination \code{lc = X'beta}}

\item{eta}{for models needing extra parameter besides beta in the linear combination \code{lc}. 
eta can be a vector}

\item{klass}{the class of model. Currently "lm" for linear regression, "logit" (logit model), 
"poisson" (poisson model), "probit" (probit model), "negbin" (negative binomial) are supported.}
}
\value{
object an object list with class attribute klass.
}
\description{
A function for constructing functions for concrete classes of models for the \code{crit} family of
of functions.
}
