% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit.lm}
\alias{crit.lm}
\title{Implemention - linear regression}
\usage{
\method{crit}{lm}(object)
}
\arguments{
\item{object}{Object with class attribute "lm" obtained from the function \code{c_crit()}}
}
\value{
sum of squared errors
}
\description{
This function evaluates the criterion for the linear regression class of model.
}
