% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit.logit}
\alias{crit.logit}
\title{Implemention - logit model}
\usage{
\method{crit}{logit}(object)
}
\arguments{
\item{object}{Object with class attribute "logit" obtained from the function \code{c_crit()}}
}
\value{
negative log likelihood
}
\description{
This function evaluates the criterion for the logit class of model.
}
