% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit.logitcd}
\alias{crit.logitcd}
\title{Implemention - logit model}
\usage{
\method{crit}{logitcd}(object)
}
\arguments{
\item{object}{Object with class attribute "logitcd" obtained from the function \code{c_crit()}}
}
\value{
negative log likelihood
}
\description{
This function evaluates the criterion for the logit class of model.
}
