% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit.nbcd}
\alias{crit.nbcd}
\title{Implemention - negative binomial model}
\usage{
\method{crit}{nbcd}(object)
}
\arguments{
\item{object}{Object with class attribute "nbcd" obtained from the function \code{c_crit()}}
}
\value{
negative log likelihood
}
\description{
This function evaluates the criterion for the negative binomial class of model.
}
