% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit.poiscd}
\alias{crit.poiscd}
\title{Implemention - poisson model}
\usage{
\method{crit}{poiscd}(object)
}
\arguments{
\item{object}{Object with class attribute "poiscd" obtained from the function \code{c_crit()}}
}
\value{
negative log likelihood
}
\description{
This function evaluates the criterion for the poisson class of model.
}
