% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit.probit}
\alias{crit.probit}
\title{Implemention - probit model}
\usage{
\method{crit}{probit}(object)
}
\arguments{
\item{object}{Object with class attribute "probit" obtained from the function \code{c_crit()}}
}
\value{
negative log likelihood
}
\description{
This function evaluates the criterion for the probit class of model.
}
