% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit}
\alias{crit}
\title{Model criterion function}
\usage{
crit(object)
}
\arguments{
\item{object}{the object to be passed to the concrete class constructor \code{c_crit()}}
}
\description{
A generic S3 function for evaluating the criteria of models, eg. sum of squares for linear 
regression, negative sum of the log of individual likelihoods for MLE, etc.
}
