% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit_obs.lmcd}
\alias{crit_obs.lmcd}
\title{Implemention - linear regression model}
\usage{
\method{crit_obs}{lmcd}(object)
}
\arguments{
\item{object}{Object with class attribute "lmcd" obtained from the function \code{c_crit()}}
}
\value{
vector of squared residuals
}
\description{
This function evaluates to a vector of squared residuals for the linear regression class of model.
}
