% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit_obs.logitcd}
\alias{crit_obs.logitcd}
\title{Implemention - logit model}
\usage{
\method{crit_obs}{logitcd}(object)
}
\arguments{
\item{object}{Object with class attribute "logitcd" obtained from the function \code{c_crit()}}
}
\value{
vector of negative log likelihoods
}
\description{
This function evaluates to a vector of log-likelihoods for the logit class of model.
}
