% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit_obs.negbin}
\alias{crit_obs.negbin}
\title{Implemention - negative binomial model}
\usage{
\method{crit_obs}{negbin}(object)
}
\arguments{
\item{object}{Object with class attribute "negbin" obtained from the function \code{c_crit()}}
}
\value{
vector of negative log likelihoods
}
\description{
This function evaluates to a vector of log-likelihoods for the negative binomial class of model.
}
