% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit_obs.pois}
\alias{crit_obs.pois}
\title{Implemention - poisson model}
\usage{
\method{crit_obs}{pois}(object)
}
\arguments{
\item{object}{Object with class attribute "pois" obtained from the function \code{c_crit()}}
}
\value{
vector of negative log likelihoods
}
\description{
This function evaluates to a vector of log-likelihoods for the poisson class of model.
}
