% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit_obs.probit}
\alias{crit_obs.probit}
\title{Implemention - probit model}
\usage{
\method{crit_obs}{probit}(object)
}
\arguments{
\item{object}{Object with class attribute "probit" obtained from the function \code{c_crit()}}
}
\value{
vector of negative log likelihoods
}
\description{
This function evaluates to a vector of log-likelihoods for the probit class of model.
}
