% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{crit_obs}
\alias{crit_obs}
\title{Model criterion function}
\usage{
crit_obs(object)
}
\arguments{
\item{object}{the object to be passed to the concrete class constructor \code{c_crit()}}
}
\description{
A generic S3 function for evaluating the criteria of models, eg. squares of residuals for linear 
regression, negative log of individual likelihoods for MLE, etc.
}
