% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cross_Dep.R
\name{gdat_cd}
\alias{gdat_cd}
\title{Generate Data}
\usage{
gdat_cd(pars, N, Tp, ncXm = 0, seed, fun, sder = NULL, eta = NULL,
  modclass = "lmcd")
}
\arguments{
\item{pars}{vector of parameters to generate the data.}

\item{N}{number of units in the panel}

\item{Tp}{number of time periods}

\item{ncXm}{number of other control variables to include}

\item{seed}{seed for random number generation}

\item{fun}{function that models cross-sectional dependence function \eqn{\phi(.,.)}}

\item{sder}{standard deviation for the error term. this option is applicable to the class "lmcd"}

\item{eta}{additional parameter, like \eqn{\theta} in the negative binomial class "nbcd" and 
seed for simulating the outcome for MLE methods. see \link{c_out_dep}}

\item{modclass}{the class of model; see \link{ncd_gen} for models supported}
}
\value{
A data frame of Y,X,psID,tpID, where psID and tpID are unit and time IDs, Xm of ncXm columns
 if ncXm is specified
}
\description{
Generate panel data with cross-sectional dependence for classes of models viz. Generate data for model 
with a single weight delta
}
\examples{
pars = c(1.0,0.5,0.8); pars2=pars = c(1.0,0.5,0.8,0.1,-0.1); N = 10; Tp = 16 
fnp<- function(x,y) {-(0.5*y^4 + (x-y)^4)^.25} #poisson model
datpois = gdat_cd(pars=pars,N=N,Tp=Tp,seed=2,fun=fnp,eta = 200,modclass="poiscd") 
datpois2 = gdat_cd(pars=pars2,N=N,Tp=Tp,ncXm=2,seed=2,fun=fnp,eta = 200,modclass="poiscd")
summary(datpois); summary(datpois2);#NA's in first period for Y due to lag
}
\seealso{
\link{c_out_dep}
}
