% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cross_Dep.R
\name{inter_xij}
\alias{inter_xij}
\title{Binary Interaction}
\usage{
inter_xij(xi, xj, k = length(xi))
}
\arguments{
\item{xi}{vector of first variable terms}

\item{xj}{vector of second variable terms}

\item{k}{a dummy input as length of xi; does not need to be specified.}
}
\value{
vector of binaries \eqn{\{0,1\}} the same length as xi
}
\description{
Interact several factors. This function allows for binary \eqn{\{0,1\}} interaction between
observable factors xi and xj. xi and xj should be of the same length. It creates a 1 if i 
and j belong to the same category (xi=xj) and zero otherwise. This function also allows 
several factors i.e. xi and xj as vectors
}
\examples{
xi=c(0,1,0); xj=c(0,1,1); inter_xij(xi,xj,2)

}
