% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cross_Dep.R
\name{ncd_gen}
\alias{ncd_gen}
\title{Cross-sectional dependence in models}
\usage{
ncd_gen(pars, Y, X, Xm = NULL, Xi = X, Tid, Pid, fun = fn4, k = 1,
  nt, utid, modclass = "lm", crit_obs = FALSE)
}
\arguments{
\item{pars}{vector of parameter values at which to evaluate the function}

\item{Y}{outcome variable}

\item{X}{matrix of covariates or design matrix}

\item{Xm}{matrix of other control variables}

\item{Xi}{matrix of variable(s) that drive cross-sectional dependence}

\item{Tid}{time ID}

\item{Pid}{unit ID}

\item{fun}{a function that parameterises the form of dependence. it can be user-written}

\item{k}{order of polynomial approximation for \code{fun()}}

\item{nt}{number of terms in approximation; should be of length equal to output of \code{fun()}}

\item{utid}{unique time IDs to be used in the estimation; should leave at least 1 period for lagged Xi}

\item{modclass}{the class of model. See description above for classes supported.}

\item{crit_obs}{a logical for vector of observation-specific criterion values, squared residuals for 
"lm", negative log-likelihoods for MLE methods, etc.}
}
\value{
criterion value or vector of observation-specific criterion values if \code{crit_obs=TRUE}
}
\description{
This routine computes all models considered for cross-sectional dependence using endogenously
generated weight matrices. Currently "lm" for linear regression, "logit" (logit model), 
"poisson" (poisson model), "probit" (probit model), "negbin" (negative binomial) are supported.
Note: the data should first be sorted by unit ID \code{Pid}, then by time ID \code{Tid}
}
