% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cross_Dep.R
\name{ncd_gen2}
\alias{ncd_gen2}
\title{Cross-sectional dependence in models}
\usage{
ncd_gen2(pars, Y, X, Xm = NULL, Tid, Pid, wfun, k = 1, nt, utid,
  modclass = "lmcd", rval = "crit", eta = 0.5, dWzero = FALSE)
}
\arguments{
\item{pars}{vector of parameter values at which to evaluate the function}

\item{Y}{outcome variable}

\item{X}{matrix of covariates that generate spillovers or cross-sectional dependence}

\item{Xm}{matrix of other control variables}

\item{Tid}{time ID}

\item{Pid}{unit ID}

\item{wfun}{a user-supplied function of indices i,j, t, and \eqn{\delta} that directly models the (i,j)'th 
\eqn{i \not = j}  entry (without normalisation) of the spatial matrix at period t. This is a more flexible option
vis-a-vis \code{fun} argument in \code{\link{ncd_gen}}. See example below.}

\item{k}{order of polynomial approximation for \code{fun()}; this can be set as a dummy for non-polynomial
forms of \code{fun}}

\item{nt}{number of terms in approximation; should be of length equal to output of \code{fun()}}

\item{utid}{unique time IDs to be used in the estimation; should leave at least 1 period for lagged Xi}

\item{modclass}{the class of model. See description above for classes supported.}

\item{rval}{character string naming the output to return; "crit" - value of criterion, 
"critobs" - observation-specific criterion values, squared residuals for 
"lmcd", negative log-likelihoods for MLE methods, etc., or "Xweight" vector X*W where W collects weight matrices
through time.}

\item{eta}{extra parameter to be passed, eg. \eqn{\theta} for negative binomial,\eqn{\tau} for quantile
regression.}

\item{dWzero}{logical. Should the diagonal elements of the weight matrix be set to zero? 
Defaults to \code{FALSE}. Choose option \code{TRUE} if own drivers \code{Xi} should affect the outcome}
}
\value{
criterion value or vector of observation-specific criterion values if \code{crit_obs=TRUE}
}
\description{
This routine computes all models considered for cross-sectional dependence using endogenously
generated weight matrices. Currently "lmcd" for linear regression, "logitcd" (logit model), 
"poiscd" (poisson model), "probitcd" (probit model), "nbcd" (negative binomial) are supported.
Note: the data should first be sorted by unit ID \code{Pid}, then by time ID \code{Tid}
}
\examples{
pars=c(1.0,0.5,0.8); N = 5; Tp = 6; fnp<- function(x,y,k){-(0.5*y^4+(x-y)^4)^.25} #dummy k
datpois=gdat_cd(pars=pars,N=N,Tp=Tp,seed=2,fun=fnp,eta = 200,modclass="poiscd") #poisson data 
k=1; lp=k*(k+1)/2; startp = rep(0.2,(lp+2)) 
wfn = function(i,j,t,del){ Xi=datpois$X; xit1 = datpois$X[((i-1)*Tp + t-1)]; 
xjt1 = datpois$X[((j-1)*Tp + t-1)]; return(exp(sum(c(fnp(xit1,xjt1,k))*del)))} #weight function
wfn(2,3,4,0.5) #test the weight function
RpacSPD::ncd_gen2(pars=startp,Y=datpois$Y,X=datpois$X,Xm=NULL,Tid=datpois$tpID,
Pid=datpois$psID,wfun=wfn,k=k,nt=lp,utid=c(2:Tp),modclass="poiscd") #return function value
RpacSPD::ncd_gen2(pars=startp,Y=datpois$Y,X=datpois$X,Xm=NULL,Tid=datpois$tpID,
Pid=datpois$psID,wfun=wfn,k=k,nt=lp,utid=c(2:Tp),modclass="poiscd",rval="Xw") #return Xw

}
