% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{out_dep.lmcd}
\alias{out_dep.lmcd}
\title{Implemention - linear regression model}
\usage{
\method{out_dep}{lmcd}(object)
}
\arguments{
\item{object}{object of class "lmcd"}
}
\value{
Y vector of outcome variable
}
\description{
Generate outcome variable for linear regression class of model
}
