% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{out_dep.logit}
\alias{out_dep.logit}
\title{Implemention - logit model}
\usage{
\method{out_dep}{logit}(object)
}
\arguments{
\item{object}{object of class "logit"}
}
\value{
Y vector of outcome variable
}
\description{
Generate outcome variable for logit regression class of model
}
