% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{out_dep.negbin}
\alias{out_dep.negbin}
\title{Implemention - negative binomial model}
\usage{
\method{out_dep}{negbin}(object)
}
\arguments{
\item{object}{object of class "negbin"}
}
\value{
Y vector of outcome variable
}
\description{
Generate outcome variable for the negative binomial class of model
}
