% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{out_dep.poiscd}
\alias{out_dep.poiscd}
\title{Implemention - poisson model}
\usage{
\method{out_dep}{poiscd}(object)
}
\arguments{
\item{object}{object of class "poiscd"}
}
\value{
Y vector of outcome variable
}
\description{
Generate outcome variable for the poisson class of model
}
