% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{out_dep.probitcd}
\alias{out_dep.probitcd}
\title{Implemention - probit model}
\usage{
\method{out_dep}{probitcd}(object)
}
\arguments{
\item{object}{object of class "probitcd"}
}
\value{
Y vector of outcome variable
}
\description{
Generate outcome variable for the probit class of model
}
