% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{out_dep}
\alias{out_dep}
\title{Randomly generate outcome Y}
\usage{
out_dep(object)
}
\arguments{
\item{object}{Object of class attribute obtained from \code{c_out_dep()}}
}
\description{
A generic function for randomly generating outcome variables for model classes.
}
