% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cross_Dep.R
\name{polyexp_mul}
\alias{polyexp_mul}
\title{Multivariate Polynomial Approximation}
\usage{
polyexp_mul(xi, xj, k)
}
\arguments{
\item{xi}{vector of first variable terms}

\item{xj}{vector of second variable terms}

\item{k}{order of polynomial approximation}
}
\value{
vector of length \eqn{l*k(k+1)/2}
}
\description{
Take a k-order polynomial approximation of a multivariate \eqn{\phi(xi,xj)} where both xi and xj 
are vectors of length l. This expansion involves l pairwise polynomial expansions using 
\code{polyexp()}. The expansion excludes terms that are solely in the first variable term 
because coefficients on them are not identified.
}
\examples{
set.seed(2); xi=stats::runif(3); xj=stats::runif(3); polyexp_mul(xi,xj,2)

}
