% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cross_Dep.R
\name{reg_cd}
\alias{reg_cd}
\title{Cross-sectional dependence in regression}
\usage{
reg_cd(startp, ..., rvcov = FALSE)
}
\arguments{
\item{startp}{vector of starting values}

\item{...}{arguments to be passed to \code{ncd_gen()} \strong{except} argument \code{rval}. 
Argument names must match exactly.}

\item{rvcov}{Logical. Should the variance-covariance matrix be returned?}
}
\value{
A list
\itemize{
  \item coefs vector of coefficients
  \item stde vector of standard errors
  \item tstat vector of t-statistics
  \item pval vector of p-values
  \item varcov variance-covariance matrix if \code{rvcov} is set to \code{TRUE}
  \item Wstat a Wald chi-square statistic
  \item pvwald Prob>Wstat 
}
}
\description{
A routine that takes starting values and model specifications as in \code{ncd_gen()} and returns
regression results.
}
\examples{
pars=c(1.0,0.5,0.8); N = 5; Tp = 6; fnp<- function(x,y,k){-(0.5*y^4+(x-y)^4)^.25} #dummy k
datpois=gdat_cd(pars=pars,N=N,Tp=Tp,seed=2,fun=fnp,eta = 200,modclass="poiscd") #poisson data 
k=1; lp=k*(k+1)/2; startp = rep(0.2,(lp+2)) # fun() is known
zg1=RpacSPD::reg_cd(startp=startp,Y=datpois$Y,X=datpois$X,Xm=NULL,Xi=datpois$X,Tid=datpois$tpID,
Pid=datpois$psID,fun=fnp,k=k,nt=lp,utid=c(2:Tp),modclass="poiscd") #return function value
BIC(zg1) #compute BIC of fitted model

k=4; lp=k*(k+1)/2; startp = rep(0.2,(lp+2)) # fun() is polynomial approximated
zg4=RpacSPD::reg_cd(startp=startp,Y=datpois$Y,X=datpois$X,Xm=NULL,Xi=datpois$X,Tid=datpois$tpID,
Pid=datpois$psID,fun=polyexp,k=k,nt=lp,utid=c(2:Tp),modclass="poiscd") #return function value
BIC(zg4) #compute BIC of fitted model
}
\seealso{
ncd_gen
}
