% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{regir.lmcd}
\alias{regir.lmcd}
\title{Regression - lmcd class}
\usage{
\method{regir}{lmcd}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and Xmat - design matrix of class "lmcd"}

\item{...}{additional parameters to be passed to \code{\link[stats]{lm}}}
}
\value{
fitted model object
}
\description{
A linear regression implementation for the "lmcd" class. It uses \code{\link[stats]{lm}} ##\code{lm()} from the stats package.
}
