% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{regir.logitcd}
\alias{regir.logitcd}
\title{Regression - logitcd class}
\usage{
\method{regir}{logitcd}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and Xmat - design matrix of class "logitcd"}

\item{...}{additional parameters to be passed to \code{\link[stats]{glm}}}
}
\value{
fitted model object
}
\description{
A logit regression implementation for the "logitcd" class. It uses \code{\link[stats]{glm}}
}
