% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{regir.poiscd}
\alias{regir.poiscd}
\title{Regression - poiscd class}
\usage{
\method{regir}{poiscd}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and Xmat - design matrix of class "poiscd"}

\item{...}{additional parameters to be passed to \code{\link[stats]{glm}}}
}
\value{
fitted model object
}
\description{
A poisson regression implementation for the "poiscd" class. It uses \code{\link[stats]{glm}}
}
