% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{regir.probitcd}
\alias{regir.probitcd}
\title{Regression - probitcd class}
\usage{
\method{regir}{probitcd}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and Xmat - design matrix of class "probitcd"}

\item{...}{additional parameters to be passed to \code{\link[stats]{glm}}}
}
\value{
fitted model object
}
\description{
A probit regression implementation for the "probitcd" class. It uses \code{\link[stats]{glm}}
}
