% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{regir.qreg}
\alias{regir.qreg}
\title{Regression - qreg class}
\usage{
\method{regir}{qreg}(object, ...)
}
\arguments{
\item{object}{a list of Y - outcome variable and Xmat - design matrix of class "qreg"}

\item{...}{additional parameters to be passed to \code{\link[quantreg]{rq}}}
}
\value{
fitted model object
}
\description{
A quantile regression implementation for the "qreg" class. It uses \code{\link[quantreg]{rq}}
}
