% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{regir}
\alias{regir}
\title{Model criterion function}
\usage{
regir(object, ...)
}
\arguments{
\item{object}{the object to be passed to the concrete class constructor \code{regir}}

\item{...}{additional paramters to be passed to the internal routine}
}
\description{
A generic S3 function as wrapper for internal R routines for classes of models implemented
in this package. See details \link{c_crit} for the list of classes supported.
}
