% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Cross_Dep.R
\name{swvcovHR}
\alias{swvcovHR}
\title{Huber-White Sandwich Estimator}
\usage{
swvcovHR(hessmat, estfmat)
}
\arguments{
\item{hessmat}{the hessian matrix of dimension kxk evaluated at the converged parameter values}

\item{estfmat}{empirical estimating function}
}
\value{
varcov a sandwich variance-covariance matrix
}
\description{
A Sandwich Huber-White variance-covariance estimator.
}
