% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_Model_Criteria.R
\name{c_out_dep}
\alias{c_out_dep}
\title{Object class constructor}
\usage{
c_out_dep(lc, eta = NULL, modclass = "lmcd")
}
\arguments{
\item{lc}{linear combination `lc = X'beta`}

\item{eta}{for models needing extra parameter besides beta in the linear combination \code{lc}. 
eta can be a vector. For "lmcd", \code{eta} is the vector of errors; must of the same length as Y
For "nbcd", \code{eta} must be a vector of two elements; the first one is parameter \code{th}
and the second is the seed for random generation. For other classes, eta is the seed for random 
generation.}

\item{modclass}{the class of model. Currently "lmcd" for linear regression, "logitcd" (logit model), 
"poiscd" (poisson model), "probitcd" (probit model), "nbcd" (negative binomial) are supported.}
}
\value{
object an object list with class attribute modclass.
}
\description{
This function constructs an object class for the generic function \code{out_dep()}
}
